Extract UnmanagedBlob to a separated file
